move enable/disable module to client
