+++ internal/kube/kyma/kyma.go
@@ -2,6 +2,7 @@ package kyma
 
 import (
 	"context"
+	"slices"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -17,6 +18,8 @@ const (
 type Interface interface {
 	GetDefaultKyma(context.Context) (*Kyma, error)
 	UpdateDefaultKyma(context.Context, *Kyma) error
+	EnableModule(context.Context, string, string) error
+	DisableModule(context.Context, string) error
 }
 
 type client struct {
@@ -29,7 +32,7 @@ func NewClient(dynamic dynamic.Interface) Interface {
 	}
 }
 
-// GetDefaultKyma uses dynamic client to get the default Kyma CR from the kyma-system namespace and cast it to the Kyma structure
+// GetDefaultKyma gets the default Kyma CR from the kyma-system namespace and cast it to the Kyma structure
 func (c *client) GetDefaultKyma(ctx context.Context) (*Kyma, error) {
 	u, err := c.dynamic.Resource(GVRKyma).
 		Namespace(defaultKymaNamespace).
@@ -44,7 +47,7 @@ func (c *client) GetDefaultKyma(ctx context.Context) (*Kyma, error) {
 	return kyma, err
 }
 
-// UpdateDefaultKyma uses dynamic client to update the default Kyma CR from the kyma-system namespace based on the Kyma CR from arguments
+// UpdateDefaultKyma updates the default Kyma CR from the kyma-system namespace based on the Kyma CR from arguments
 func (c *client) UpdateDefaultKyma(ctx context.Context, obj *Kyma) error {
 	u, err := runtime.DefaultUnstructuredConverter.ToUnstructured(obj)
 	if err != nil {
@@ -56,4 +59,54 @@ func (c *client) UpdateDefaultKyma(ctx context.Context, obj *Kyma) error {
 		Update(ctx, &unstructured.Unstructured{Object: u}, metav1.UpdateOptions{})
 
 	return err
+}
+
+// EnableModule adds module to the default Kyma CR in the kyma-system namespace
+// if moduleChannel is empty it uses default channel in the Kyma CR
+func (c *client) EnableModule(ctx context.Context, moduleName, moduleChannel string) error {
+	kymaCR, err := c.GetDefaultKyma(ctx)
+	if err != nil {
+		return err
+	}
+
+	kymaCR = enableModule(kymaCR, moduleName, moduleChannel)
+
+	return c.UpdateDefaultKyma(ctx, kymaCR)
+}
+
+// DisableModule removes module from the default Kyma CR in the kyma-system namespace
+func (c *client) DisableModule(ctx context.Context, moduleName string) error {
+	kymaCR, err := c.GetDefaultKyma(ctx)
+	if err != nil {
+		return err
+	}
+
+	kymaCR = disableModule(kymaCR, moduleName)
+
+	return c.UpdateDefaultKyma(ctx, kymaCR)
+}
+
+func enableModule(kymaCR *Kyma, moduleName, moduleChannel string) *Kyma {
+	for i, m := range kymaCR.Spec.Modules {
+		if m.Name == moduleName {
+			// module already exists, update channel
+			kymaCR.Spec.Modules[i].Channel = moduleChannel
+			return kymaCR
+		}
+	}
+
+	kymaCR.Spec.Modules = append(kymaCR.Spec.Modules, Module{
+		Name:    moduleName,
+		Channel: moduleChannel,
+	})
+
+	return kymaCR
+}
+
+func disableModule(kymaCR *Kyma, moduleName string) *Kyma {
+	kymaCR.Spec.Modules = slices.DeleteFunc(kymaCR.Spec.Modules, func(m Module) bool {
+		return m.Name == moduleName
+	})
+
+	return kymaCR
 }
